ARM,crt0.S: optional init gd to gdata for spl
The common/spl changes gd to gdata in board_init_f.
I fail to see why. The only reason I can think of to
use the gdata is the case where is won't fit into the
region where the stack lives.

Move the init to crt0.S and make it a CONFIG.

Conflicts:
	arch/arm/lib/crt0.S